% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddCruiseBehavior}
\alias{AddCruiseBehavior}
\title{Adds cruise behavior}
\usage{
AddCruiseBehavior(df, min_agl, min_speed, threshold_agl)
}
\arguments{
\item{df}{dataframe}

\item{min_agl}{distance (in meters) above ground}

\item{min_speed}{minimum speed of bird}

\item{threshold_agl}{any locations above this threshold of agl ('above ground
level') are labeled "cruise"}
}
\value{
dataframe with behavior column that has "cruise"
}
\description{
Finds cruise behavior that meet given threshold parameters
}
\details{
should run AddNestBehavior, AddRoostBehavior, and AddCruiseBehavior
  prior to this function
}
