% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddFlightBehavior}
\alias{AddFlightBehavior}
\title{Adds flight behavior}
\usage{
AddFlightBehavior(df, min_speed, min_step_length, max_step_time,
    threshold_agl)
}
\arguments{
\item{df}{dataframe}

\item{min_speed}{minimum speed of bird}

\item{min_step_length}{distance (in meters) between locations}

\item{max_step_time}{maximum time between locations}

\item{threshold_agl}{any locations above this threshold of agl ('above ground
level') are labeled "cruise"}
}
\value{
dataframe with columns for flight_index, flight_step, and
 flight_length
}
\description{
Finds location data that meet given threshold parameters for flights, which
  include the start point, mid-flight, and final location of a flight.
}
\details{
adds 'bh_flight' to dataframe
}
