% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddFlightData}
\alias{AddFlightData}
\title{Adds flight behavior}
\usage{
AddFlightData(df, min_speed = 5, min_step_length = 50, max_step_time = 20)
}
\arguments{
\item{df}{dataframe}

\item{min_speed}{minimum speed of bird}

\item{min_step_length}{distance (in meters) between locations}

\item{max_step_time}{maximum time between locations}
}
\value{
dataframe with columns for flight_index, flight_step, and
 flight_length
}
\description{
Finds location data that meet given threshold parameters for flights, which
  include the start point, mid-flight, and final location of a flight.
}
\details{
should run AddLandscapeValues() prior to this function
}
