% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddHomeConEdgeDistanceBAEA}
\alias{AddHomeConEdgeDistanceBAEA}
\title{Adds home and conspecific edge distances to 'baea' dataframe}
\usage{
AddHomeConEdgeDistanceBAEA(baea, nest_set, base, output_dir = getwd(), max_r,
  write_home_dist = FALSE, write_con_dist = FALSE,
  write_con_dist_nest = FALSE, write_edge_dist = FALSE,
  write_terr_edge = FALSE)
}
\arguments{
\item{baea}{dataframe of baea locations}

\item{nest_set}{dataframe of nests}

\item{base}{base Raster that sets the projection, extent, and dimensions of
the study area}

\item{output_dir}{directory for output files (distance, homerange)}

\item{max_r}{maximum radius to calculate the homerange raster from each
df_home centroid}

\item{write_home_dist}{logical, write home nest distance raster to file.
Default is FALSE.}

\item{write_con_dist}{logical, write conspecific nest distance raster to
file. Default is FALSE.}

\item{write_con_dist_nest}{logical, write conspecific nest distance with all
values centered on zero at the nest as raster to file. Default is FALSE.}

\item{write_edge_dist}{logical, write territorial edge distance raster to
file. Default is FALSE.}

\item{write_terr_edge}{logical, write territorial edge raster to file.
Default is FALSE.}
}
\value{
baea
}
\description{
Creates Raster Layers of homerange kernels based on home centroids
}
\details{
Creates folders for every year in output directory
}
