% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddHomeRangeKernelsBAEA}
\alias{AddHomeRangeKernelsBAEA}
\title{Creates and exports homerange kernels}
\usage{
AddHomeRangeKernelsBAEA(baea, nest_set, set_name, base, output_dir,
  max_r, home_inflection, home_scale, avoid_inflection, avoid_scale,
  min_prob, write_distance, write_homerange)
}
\arguments{
\item{baea}{dataframe of all homerange centroids}

\item{nest_set}{dataframe of nests}

\item{set_name}{character, name of nest set}

\item{base}{base Raster that sets the projection, extent, and dimensions of
the study area}

\item{output_dir}{directory for output files (distance, homerange)}

\item{max_r}{maximum radius to calculate the homerange raster from each
df_home centroid}

\item{home_inflection}{inflection point of the Logistic function that governs
the home kernel}

\item{home_scale}{scale parameter of the Logistic function that governs the
scale parameter}

\item{avoid_inflection}{inflection point of the Logistic function that
governs the conspecific avoidance kernel}

\item{avoid_scale}{scale parameter of the Logistic function that governs
the conspecific avoidance kernel}

\item{min_prob}{numeric, minimum probablity threshold; all probabilities
below this value will be rounded to zero. Default is 0.}

\item{write_distance}{logical, write distance raster to file. Default is
FALSE.}

\item{write_homerange}{logical, write home range raster to file. Default is
FALSE.}
}
\value{
A list containing homerange kernel Rasters for all the df_home
  centroids
}
\description{
Creates RasterLayers of homerange kernels based on home centroids
}
\details{
Create a folder for every year in the output directory
}
