% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddNestBehavior}
\alias{AddNestBehavior}
\title{Adds nest behavior to baea dataframe}
\usage{
AddNestBehavior(df = df, distance_threshold = 50)
}
\arguments{
\item{df}{dataframe}

\item{distance_threshold}{distance (in meters) from location to nest to assign
a location as "nest" behavior}
}
\value{
dataframe with 'bh_nest' column that has "nest"
}
\description{
Finds nest attendance behavior that meet given threshold parameters
}
\details{
need to run AddNestData() prior to running this
}
