% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddNestConDist}
\alias{AddNestConDist}
\title{Adds nest and conspecific distances}
\usage{
AddNestConDist(df, nests_con_dist)
}
\arguments{
\item{df}{dataframe of locations, must have coordinate columns}
}
\value{
dataframe with "nest_dist" and "con_dist" columns
}
\description{
Adds nest and conspecific distances to dataframe
}
\details{
The nest_con_dist has to have structure of list[[year]][[nest_id]]
}
