% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{AddNormDens}
\alias{AddNormDens}
\title{Adds Normal Distribution Density}
\usage{
AddNormDens(x)
}
\arguments{
\item{Numeric}{vector}
}
\value{
A vector
}
\description{
Returns a vector of normal distribution density values based on the input
  values and input data's estimated parameters.
}
\details{
Designed for use in plyr transform functions
}
