% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddPerchBehavior}
\alias{AddPerchBehavior}
\title{Add perch behavior}
\usage{
AddPerchBehavior(df, max_speed)
}
\arguments{
\item{df}{dataframe}

\item{max_speed}{numeric, speed over which behavior is no longer considered
perch.}
}
\value{
dataframe with behavior column that has "perch"
}
\description{
Finds perch behavior that meets given threshold parameters
}
