% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddRoostBehavior}
\alias{AddRoostBehavior}
\title{Adds roost behavior to dataframe}
\usage{
AddRoostBehavior(df, at_roost_distance_threshold, depart_timediff_max,
  arrive_timediff_max)
}
\arguments{
\item{df}{dataframe}

\item{at_roost_distance_threshold}{numeric max distance away from roost}

\item{depart_timediff_max}{max diff between start of day and depart}

\item{arrive_timediff_max}{max diff between end of day and arrival}
}
\value{
dataframe with 'bh_roost' column that has arrive, depart, and roost
}
\description{
Finds roost arrivals and departures that meet given threshold parameters
}
\details{
Adds 'bh_roost' column
}
