% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddSeason}
\alias{AddSeason}
\title{Adds 'season' to dataframe}
\usage{
AddSeason(df, date_col)
}
\arguments{
\item{df}{dataframe}

\item{by}{date column. Default is "date".}
}
\value{
dataframe with 'season' column
}
\description{
Adds 'season' column to original dataframe. Seasons start at 3-19, 6-20,
  9-21, and 12-20 for "Spring", "Summer", "Fall", and "Winter", respectively.
}
