% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddTimeStepProportion}
\alias{AddTimeStepProportion}
\title{Adds time step proportions}
\usage{
AddTimeStepProportion(
  df = df,
  by = "id",
  time_step = "15 min",
  tz = "Etc/GMT+5"
)
}
\arguments{
\item{df}{dataframe}

\item{by}{grouping column. Default is "id"}

\item{time_step}{time step length, based on lubriate times. Default is
"15 min".}

\item{tz}{timezone for dataset. Default is "Etc/GMT+5"}
}
\value{
dataframe with time_steps, day_min, time_after_start, and
  time_proportion data
}
\description{
Adds time_steps, day_min, time_after_start, and time_proportion data
}
\details{
need to run AddSolarTimes() prior to running this function
}
