% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AnalyzePairLocations}
\alias{AnalyzePairLocations}
\title{Analyses overlap of homeranges in paired nest locations}
\usage{
AnalyzePairLocation(pair, nests, points, mid_length, zoom)
}
\arguments{
\item{pair}{pair to be analyzed, first nest is analyzed}

\item{nests}{nest locations}

\item{points}{baea locations}

\item{mid_length}{length of mid-line.}

\item{zoom}{integer (2-22) of google map zoom, default = 10.}
}
\value{
Multiple output: writes .csv, plots, and dataframe
}
\description{
Analyzes pair locations to determine mid-way point betweeen nests and
  proportion of overlapping locations
}
