% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{BehaviorSubModelBAEA}
\alias{BehaviorSubModelBAEA}
\title{Submodel for behavioral transisitions in BAEA IBM}
\usage{
BehaviorSubModelBAEA(
  sim = sim,
  agent_states = agent_states,
  step_data = step_data,
  step = step
)
}
\arguments{
\item{step}{}
}
\value{

}
\description{
Submodel for behavioral transisitions in BAEA IBM
}
