% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Bin1DimByGroup}
\alias{Bin1DimByGroup}
\title{Bin one dimension vector by group}
\usage{
Bin1DimByGroup(df, var, group, breaks, style, n, midpoints, melt,
  signif, ...)
}
\arguments{
\item{df}{Dataframe with 2 columns.}

\item{var}{Column name of variable to bin}

\item{group}{Column name of group}

\item{breaks}{Numeric, vector of bin ranges}

\item{style}{String, any "breaks" type compatible with classInterval().
Default is "pretty".}

\item{n}{Numeric, number of bins, as used by classInterval(). Default is 5.}

\item{midpoints}{Logical, whether to use midpoints for labels. Otherwise,
ranges are used. Default is TRUE.}

\item{melt}{Logical, whether not to melt() the data}

\item{signif}{Integer, the number of digits used in formatting the break
numbers. Default is 5.}

\item{...}{Arguments to be passed to the functions called in each style used
by classIntervals()}
}
\value{
Two option: If melt = FALSE, returns the original dataframe with
  mid-points or ranges for the variable (with suffixes "_mid" or "_bin"). If
  melt = TRUE, a table of frequencies ("freq"), and densitities ("dens") for
  each group if "group" parameter is used.
}
\description{
Creates a "melted" dataframe of frequencies and probabilities for binned
  values
}
