% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Bin2Dim}
\alias{Bin2Dim}
\title{Bin two-dimension vectors}
\usage{
Bin2Dim(df, breaks1, breaks2, midpoints, include_lowest, signif)
}
\arguments{
\item{df}{Dataframe of 2 columns}

\item{breaks1}{Vector of bin ranges or any "breaks" type compatible with
hist(), used on x[,1]. Default = "Sturges"}

\item{breaks2}{Vector of bin ranges or any "breaks" type compatible with
hist(), used on x[,2]. Default = "Sturges"}

\item{midpoints}{Logical, whether or not to use midpoints for labels.
Otherwise, ranges are used. Default = TRUE.}

\item{include_lowest}{Logical, indicating if an value equal to the lowest
'breaks' value should be included.}

\item{signif}{Integer, the number of digits used in formatting the break
numbers. Default = 5.}
}
\value{
a dataframe of binned mid-points or ranges, frequencies ("freq"),
  and probabilities ("prob")
}
\description{
Creates a "melted" dataframe of frequencies and probabilities for binned
  values in 2 arrays
}
