% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ConvertNestNumToId}
\alias{ConvertNestNumToId}
\title{Converts nest numeric id to nest alphanumeric id}
\usage{
ConvertNestIdToNum(df)
}
\arguments{
\item{df}{input dataframe with "nest_id_num" column}
}
\value{
A dataframe with a numberic "nest_id" column
}
\description{
Converts a numeric "nest_id_num" column to an alphanumeric "nest_id" column
}
\details{
If a "nest_id" column doesn't exist, the function automatically
  creates one
}
