% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{CreateColorsByAny}
\alias{CreateColorsByAny}
\title{Creates colors based on any factor in the dataframe}
\usage{
CreateColorsByAny(by, df, r_pal, b_pal, output, plot, ...)
}
\arguments{
\item{by}{variable to determine colors. Specific outcomes for: "behavior",
"id", or "sex"}

\item{df}{dataframe with "by" variable - only required if "by" is not
"behavior", "id", or "sex"}

\item{pal}{dataframe with "by" variable - only required if "by" is not
"behavior", "id", or "sex"}

\item{r_pal}{color palette for CreateVarsColors(), default is NULL}

\item{b_pal}{color palette from RColorBrewer for CreateVarsColors(), default
is "Set1".}

\item{output}{color palette from RColorBrewer for CreateVarsColors(),
default is "Set1"}

\item{plot}{logical, whether or not to display names and colors, default is
FALSE}

\item{...}{additional parameters for CreateColorsByMetadata()}
}
\value{
df with "by" variable and hexidecimal colors
}
\description{
Creates and/or displays dataframe of "by" variable and associated colors#'
}
\details{
Used in several other functions. Color palettes determined
  automatically for "behavior", "id", and "sex". Others set by pal or b_pal.
  Requires 'kml' associated functions.
}
