% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{CreateConNestDistRasters}
\alias{CreateConNestDistRasters}
\title{Creates conspecific and nest distance raster for each baea in data}
\usage{
CreateConNestDistRasters(baea, nest_set, base = base,
  output_dir = "Output/Analysis/Territorial", max_r = 30000,
  write_con_nest_all = TRUE)
}
\arguments{
\item{baea}{dataframe of baea locations}

\item{nest_set}{dataframe of nests}

\item{base}{base Raster that sets the projection, extent, and dimensions of
the study area}

\item{output_dir}{directory for output files (distance, homerange)}

\item{max_r}{maximum radius to calculate the homerange raster from each
df_home centroid}

\item{write_con_nest_all}{logical, write con_nest_all raster to file.
Default is TRUE}
}
\value{
RasterLayer
}
\description{
Creates conspecific and nest distance raster for each baea in data
}
\details{
Creates folders for every year in output directory
}
