% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{CreateFlightPathSegments}
\alias{CreateFlightPathSegments}
\title{Create a dataframe of flight path segments}
\usage{
CreateFlightPathSegments(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with columns for 'id', 'behavior', and 'path_seg'.
}
\description{
Adds flight path segments, which includes locations before and after "flight"
behavior. Segments are sequentially numbered for each individual. Used for
visualization and analysis of flight paths. Some locations may be represented
twice because they are both the start and end of a path.
}
\details{
adds 'path_seg' to dataframe
}
