% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{CreateHomeRangeKernelsInterpolate}
\alias{CreateHomeRangeKernelsInterpolate}
\title{Create homerange kernels based on interpolating empiricial x,y location data}
\usage{
CreateHomeRangeKernelsInterpolate(df_all, df_home, base, max_r, id,
  name, min_prob, output_dir, write_homerange)
}
\arguments{
\item{df_all}{dataframe of all homerange centroids}

\item{df_home}{dataframe of homerange centroids to calculate homerange
kernels (these must be a subset of the df_all dataframe), Default is to
use 'df_all' dataframe}

\item{base}{base Raster that sets the projection, extent, and dimensions of
the study area}

\item{max_r}{maximum radius to calculate the homerange raster from each
df_home centroid}

\item{id}{column name of df_home that identifies the homerange.}

\item{name}{column name of df_home that identifies the name of the homerange
and is used to write the name of the .tif file.  Default is 'id', which
sets the names to df_home row number if 'id' is NULL.}

\item{min_prob}{numeric, minimum probablity threshold; all probabilities
below this value will be rounded to zero.  Default is 0.}

\item{output_dir}{directory for output files (distance, homerange)}

\item{write_distance}{logical, write distance range raster to file. Default
is FALSE}

\item{write_homerange}{logical, write home range raster to file. Default is
FALSE}
}
\value{
A list containing homerange kernel Rasters for all the df_home
  centroids
}
\description{
Creates RasterLayers of homerange kernels based on home centroids
}
\details{
Used inside AddHomeRangeKernelsBAEA()
}
