% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{CreateListIndex}
\alias{CreateListIndex}
\title{Create list index}
\usage{
CreateListIndex(list)
}
\arguments{
\item{list}{List of input data}
}
\value{
List with sequential numbers for all of the values
}
\description{
Creates a "indexed" list by changing all the values into sequential numbers
}
\details{
Works with up to 3 nested levels (lists, sublists, subssublists)
}
