% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{CreateNewPars}
\alias{CreateNewPars}
\title{Create new parameters list}
\usage{
CreateNewPars(index, random)
}
\arguments{
\item{index}{Logical, whether or not the parameter values should be numbers
indicating their relative position. Default is TRUE.}

\item{random}{Logical, whether or not the parameter values should be filled
in using runif(1,1,2), random=TRUE overrides index=TRUE. Default is FALSE.}
}
\value{
A list of newly generated behavior parameters with NA or index
  numbers for each of the values
}
\description{
Creates a new list of behavior parameters with either NA or index values.
}
\details{
Used to create a new behavior parameters list from scratch. Uses
  ExtractParsIndex().
}
