% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{CutProportion}
\alias{CutProportion}
\title{Convert data into factors}
\usage{
CutProportion(data, breaks = breaks)
}
\arguments{
\item{data}{numeric values}

\item{breaks}{number of break groups}
}
\value{
factor vector
}
\description{
Convert data into factors based on assignment of proportional groups
dicatated by breaks
}
