% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{CutProportionMid}
\alias{CutProportionMid}
\title{Convert data into midpointS}
\usage{
CutProportionMid(data, breaks = breaks)
}
\arguments{
\item{breaks}{}
}
\value{
vector
}
\description{
Convert data into midpoint values based on assignment of proportional groups
dicatated by breaks
}
