% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{DensAsymGenVonMises}
\alias{DensAsymGenVonMises}
\title{Density of asymetric generalized von Mises}
\usage{
DensAsymGenVonMises(x, mu1, mu2, kappa1, kappa2)
}
\arguments{
\item{x}{Vector}

\item{mu1}{Numeric, primary direction parameter}

\item{mu2}{Numeric, secondary direction parameter}

\item{kappa1}{Numeric (non-negative), parameter}

\item{kappa2}{Numeric (non-negative), parameter}
}
\value{
NLL of general von Mises distribution
}
\description{
Density of an asymetric generalized von Mises distribution. WORK IN
  PROGRESS!
}
