% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{DensGenVonMises}
\alias{DensGenVonMises}
\title{Density of asymetric generalized von Mises}
\usage{
DensGenVonMises(data, mu1, mu2, kappa1, kappa2)
}
\arguments{
\item{x}{Vector}

\item{mu1}{Primary direction parameter}

\item{mu2}{Secondary direction parameter}

\item{kappa1}{Numeric (non-negative), parameter}

\item{kappa2}{Numeric (non-negative), parameter}
}
\value{
NLL of general von Mises distribution
}
\description{
Density function of a generalized von Mises probability distribtion
}
\details{
Used in NLL_GenVonMises()
}
