% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExportTelemetryDataById}
\alias{ExportTelemetryDataById}
\title{Exports telemetry data into 'Individuals' folders}
\usage{
ExportTelemetryDataById(data, output_dir = file.path("C:/Work/R/Data",
  "BAEA/Telemetry/Individuals/CSVs"),
  output_dir_all = file.path("C:/Work/R/Data", "BAEA/Telemetry/All/CSVs"),
  tz = "Etc/GMT+5")
}
\arguments{
\item{data}{Dataframe of location data}

\item{output_dir}{String, folder for 'individual' files. Default is:
"C:/Work/R/Data/BAEA/Telemetry/Individuals/CSVs"}

\item{output_dir_all}{String, folder for 'all' files. Default is:
"C:/Work/R/Data/BAEA/Telemetry/Individuals/CSVs"}

\item{tz}{String, time zone. Default is "Etc/GMT+5"}
}
\value{
Exports .csv files.
}
\description{
Splits data by ID and exports .csv files to the Individuals folders
}
