% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExportTelemetryKMLByYear}
\alias{ExportTelemetryKMLByYear}
\title{Exports kml files for each year}
\usage{
ExportTelemetryKMLByYear(data, output_dir, update, tz)
}
\arguments{
\item{data}{Dataframe of location data.}

\item{output_dir}{Output directory, default is:
"C:/Work/R/Data/BAEA/Telemetry/Individuals/CSVs"}

\item{update}{Logical, whether or not to update only the current year's
files. Default is TRUE.}

\item{tz}{String, timezone of data. Default is "Etc/GMT+5".}
}
\value{
Exports .kml files
}
\description{
Splits data by ID and exports .kmls to the Individuals folders
}
