% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExtractFlightSpeed}
\alias{ExtractFlightSpeed}
\title{Extracts flight data}
\usage{
ExtractFlightData(df)
}
\arguments{
\item{df}{String, dataframe with locations.}

\item{min_speed}{Numeric, minimum speed to be classified as flight.}
}
\value{
A dataframe of "flight" data
}
\description{
Extracts the data associated with 'flight' for each bird, based on speed
  data.
}
\examples{

}
