% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExtractMovements}
\alias{ExtractMovements}
\title{Extract movement data}
\usage{
ExtractMovements(df, by, min_step_length, max_step_length)
}
\arguments{
\item{df}{Dataframe with location data which includes "step_length" and
"step_time" columns.}

\item{by}{String, column name of unique identifier to split data, default is
"id".}

\item{min_step_length}{Numeric, threshold distance for movement
classification. Default is 50.}

\item{max_step_time}{Numeric, threshold difference in time for movement
classification. Default is 20.}
}
\value{
A dataframe of selected movements
}
\description{
Extracts detected movements (non-stationary between locations) data for each
  bird.
}
