% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ExtractParsBySex}
\alias{ExtractParsBySex}
\title{Extract parameters by sex}
\usage{
ExtractParsBySex(list, sex = NULL)
}
\arguments{
\item{list}{List of parameters}

\item{sex}{String, sex to return: "male", "female", or NULL(both), default
is NULL.}
}
\value{
Dataframe of behavior parameters
}
\description{
Extracts and returns a dataframe of parameter values by sex
}
