% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ExtractParsMatrix}
\alias{ExtractParsMatrix}
\title{Extract parameters matrix}
\usage{
ExtractParsMatrix(pars)
}
\arguments{
\item{list}{List of parameters}
}
\value{
Dataframe of behavior parameters
}
\description{
Extracts a dataframe of parameter values with columns for "sex" and
  "behavior"
}
\details{
Useful for comparing parameter values among behaviors
}
