% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExtractRoostData}
\alias{ExtractRoostData}
\title{Extracts roost data}
\usage{
ExtractRoostData(df)
}
\arguments{
\item{df}{Dataframe of location data with "sex", "datetime", "behavior",
"hr_after_sunset", and "hr_before_sunrise" columns.}
}
\value{
a dataframe
}
\description{
Extracts dataframe of sex, arr_diff_min, and dep_diff_min
}
\details{
Used in other functions.
}
