% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{FilterByNestCriteria}
\alias{FilterByNestCriteria}
\title{Filter Locations by Nest Behavior Criteria}
\usage{
FilterByNestCriteria(df, min_daily_nest_dist, roll_days,
  min_roll_days_nest_dist, seasons)
}
\arguments{
\item{df}{dataframe of locations}

\item{min_daily_nest_dist}{numeric, bird's daily minimum distance from nest
to meet the criteria}

\item{roll_days}{numberic, number of days for rolling mean calculation of
nest distance}

\item{min_roll_days_nest_dist}{numeric, bird's daily minimum distance from
nest to meet the criteria}

\item{seasons}{seasons to keep (e.g., c("spring","summer")).}
}
\value{
dataframe
}
\description{
Filter Locations by Nest Behavior Criteria
}
