% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{FilterByRoostCriteria}
\alias{FilterByRoostCriteria}
\title{Filter Locations by Roost Criteria}
\usage{
FilterByRoostBehavior(df, min_daily_nest_dist, roll_days,
  min_roll_days_nest_dist, seasons)
}
\arguments{
\item{df}{dataframe of locations}

\item{overnight_distance_threshold}{numeric, distance that last PM location
and first AM location must be less than for it to be considered a
confirmed roost location}

\item{number_pm_loc}{numeric, number of GPS locations needed for PM window,
default is 7.}

\item{number_am_loc}{numeric, number of GPS locations needed for AM window,
default is 7.}

\item{tz}{timezone, default is "Etc/GMT+5"}
}
\value{
dataframe
}
\description{
Filter Locations by Roost Criteria
}
