% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{FilterLocations}
\alias{FilterLocations}
\title{Filter locations by individual and dates}
\usage{
FilterLocations(df = df, id = "id", individual = NULL, start = NULL,
  end = NULL, behavior = NULL)
}
\arguments{
\item{df}{Dataframe with locations.}

\item{id}{Column name of unique identifier. Default is "id".}

\item{individual}{String, individual/s (from id column) to keep, format
should be c(id, id), default is to keep all.}

\item{start}{Start date filter, default is 1970-01-01.}

\item{end}{End date filter, default is current date.}

\item{behavior}{String of specific behaviors to maintain. Default is all.}
}
\value{
A dataframe with subsetted rows
}
\description{
Used to filter full dataset to individual(s) within a specified date range.
}
\details{
Defaults are specific to my file directories and locations
}
