% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitArrival}
\alias{FitArrival}
\title{Fits arrival parameter}
\usage{
FitArrival(data, shape)
}
\arguments{
\item{data}{Dataframe with a "datetime", "hr_after_sunset" and "behavior"
column}

\item{shape}{Numeric, starting value of shape, default is 10.}
}
\value{
List of arr_scale, arr_shape
}
\description{
Fits Weibull parameters to roost arrival data
}
