% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitBehavior}
\alias{FitBehavior}
\title{Fits behavior parameter}
\usage{
FitBehavior(df, shape1, shape2)
}
\arguments{
\item{data}{Dataframe with a "datetime", "hr_after_sunset" and "behavior"
column}

\item{shape1}{Starting value of shape1, default is 1.}

\item{shape2}{Starting value of shape2, default is 1.}
}
\value{
List of (shape1, shape2)
}
\description{
Fit Beta parameters to behavior data
}
