% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitDeparture}
\alias{FitDeparture}
\title{Fits departure parameters}
\usage{
FitDeparture(data, shape)
}
\arguments{
\item{data}{Dataframe with a "datetime", "hr_before_sunrise", and "behavior"
column}

\item{shape}{Numeric, starting value of shape. Default is 10.}
}
\value{
List of (dep_scale, dep_shape)
}
\description{
Fit Weibull parameters to roost departure data
}
\examples{

}
