% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitGammaParsToData}
\alias{FitGammaParsToData}
\title{Fits Gamma parameter}
\usage{
FitGammaParsToData(df, var, by = NULL, scale = 1, shape = 1)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Numeric, variable to fit Gamma distribution}

\item{by}{Column name to subset data}

\item{scale}{Numeric, starting value of scale, default is 1}

\item{shape}{Numeric, starting value of shape, default is 1}
}
\value{
Dataframe of "by" variable and parameters
}
\description{
Fits a Gamma distribution to data
}
