% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitGenVonMisesDensityToArray}
\alias{FitGenVonMisesDensityToArray}
\title{Fit generalized Von Mises density to array}
\usage{
FitGenVonMisesDensityToArray(var, pars)
}
\arguments{
\item{var}{Variable fitted with generalized von Mises distribution}

\item{pars}{Parameters of generalized von Mises distribution, created by
FitGenVonMisesParsToData()}
}
\value{
Probability densities of a generalized von Mises distribution along
  an array of values
}
\description{
Fit generalized von Mises probability distribution to an array of x values
}
\details{
The x-axis is set to have 359 values, starting with 1
}
