% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitGenVonMisesParsToData}
\alias{FitGenVonMisesParsToData}
\title{Fit generalized von Mises parameter to data}
\usage{
FitGenVonMisesParsToData(df, by, var, mu1, mu2, kappa1, kappa2)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Variable to fit generalized von Mises distribution}

\item{by}{Column to subset data}

\item{mu1}{Primary direction parameter}

\item{mu2}{Secondary direction parameter}

\item{kappa1}{Numeric (non-negative) parameter of the distribution}

\item{kappa2}{Numeric (non-negative) parameter of the distribution}
}
\value{
Dataframe of "by" variable and parameters
}
\description{
Fits a generalized von Mises distribution to data
}
