% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitNormalDensityToArray}
\alias{FitNormalDensityToArray}
\title{Fits a Normal density to an array}
\usage{
FitNormalDensityToArray(df, var, pars, xlim = NULL)
}
\arguments{
\item{df}{Dataframe with data}

\item{var}{Column name of variable fitted with normal distribution}

\item{pars}{Parameters of normal distribution, created by
FitNormalParsToData()}

\item{xlim}{Numeric, sets x-axis limit max, default is: max(df[,var])}
}
\value{
Probability densities of a normal distribution along an array of
  values
}
\description{
Fit Normal Probability density function to an array of x values
}
\details{
The x-axis is set to have 500 values, starting with: min(df[,
  var])+1
}
