% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitNormalParsToData}
\alias{FitNormalParsToData}
\title{Fits normal parameters to data}
\usage{
FitNormalParsToData(df, var, by)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Column name of variable to fit Wrapped Normal distribution}

\item{by}{Column name used to subset data}
}
\value{
Dataframe of "by" variable and parameters
}
\description{
Fits a normal distribution to data
}
