% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitParetoDensityToArray}
\alias{FitParetoDensityToArray}
\title{Fit Pareto density to array}
\usage{
FitParetoDensityToArray(df, var, pars, xlim)
}
\arguments{
\item{df}{Dataframe with data}

\item{var}{Column name of variable fitted with Pareto distribution}

\item{pars}{Parameters of Pareto distribution, created by
FitParetoParsToData()}

\item{xlim}{Numeric, x-axis limit max, default is: max(df[,var])}
}
\value{
Probability densities of a Pareto distribution along an array of
  values
}
\description{
Fit Pareto probability density function to an array of x values
}
\details{
The x-axis is set to have 500 values, starting with: min(df[,
  var])+1
}
