% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitParetoParsToData}
\alias{FitParetoParsToData}
\title{Fits Pareto parameters to data}
\usage{
FitParetoParsToData(df, var, by, location, scale, shape)
}
\arguments{
\item{df}{Dataframe of input data}

\item{var}{Column name of variable to fit Pareto distribution}

\item{by}{Column name to subset data}

\item{location}{Numeric, starting value of location, default is:
min(df$var)-1.}

\item{scale}{Numeric, starting value of scale, default is 1.}

\item{shape}{Numeric, starting value of shape, default is 0.}
}
\value{
Dataframe of "by" variable and parameters
}
\description{
Fits a generalized Pareto distribution to data
}
