% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitRoost}
\alias{FitRoost}
\title{Fits roost parameters}
\usage{
FitRoost(df, pars)
}
\arguments{
\item{df}{Dataframe of location data}

\item{pars}{Input parameters to have roost pars appended onto}
}
\value{
List of parameters
}
\description{
Fits roost arrival and departure Weibull functions
}
\details{
Uses ArriveFit() and DepartFit()
}
