% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitWrappedCauchyDensityToArray}
\alias{FitWrappedCauchyDensityToArray}
\title{Fit wrapped Cauchy density to array}
\usage{
FitWrappedCauchyDensityToArray(df, var, pars)
}
\arguments{
\item{var}{Column name of variable fitted with Wrapped Cauchy distribution}

\item{pars}{Parameters of Wrapped Cauchy distribution, created by
FitWrappedCauchyToData()}
}
\value{
Probability densities of a Wrapped Cauchy distribution along an
  array of values
}
\description{
Fit wrapped Cauchy probability density function to an array of x values
}
\details{
The x-axis is set to have 359 values, starting with 1
}
