% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitWrappedCauchyParsToData}
\alias{FitWrappedCauchyParsToData}
\title{Fits a wrapped Cauchy distribution}
\usage{
FitWrappedCauchyParsToData(df, var, by = NULL, mu = 1, rho = 0)
}
\arguments{
\item{df}{Dataframe of input data}

\item{var}{Column name of variable to fit Wrapped Cauchy distribution}

\item{by}{Column name with factor to subset data}

\item{mu}{Numeric, starting value of mu, default is 1}

\item{rho}{Numeric, starting value of rho, default is 0}
}
\value{
Dataframe of "by" variable and parameters
}
\description{
Fits a wrapped Cauchy distribution to data
}
