% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FitWrappedNormalDensityToArray}
\alias{FitWrappedNormalDensityToArray}
\title{Fit wrapped Normal density to array}
\usage{
FitWrappedNormalDensityToArray(var, pars)
}
\arguments{
\item{var}{Column name of variable fitted with Wrapped Normal distribution}

\item{pars}{Parameters of Wrapped Normal distribution, created by
FitWrappedNormalToData()}
}
\value{
The probability densities of a Wrapped Normal distribution along an
  array of values
}
\description{
Fit wrapped Normal probability density function to an array of x-values
}
\details{
FitWrappedNormalDensityToArray(var, pars)

The x-axis is set to have 359 values, starting with 1
}
