% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{IfElseTimedateCompare}
\alias{IfElseTimedateCompare}
\title{Compares two times and returns greater or lesser value}
\usage{
IfElseTimedateCompare(df, col1, col2, result, default_tz, tz)
}
\arguments{
\item{df}{Dataframe of location data.}

\item{col1}{Column name, col1 is compared to col2}

\item{sign}{String, either ">" or "<" for returning greater than or less than
values, respectively. Default is ">".}

\item{col2}{Column name, col2 is compared to col1.}

\item{result}{String, name of column to insert results. Default is "result".}

\item{default_tz}{String, timezone that timedate reverts to, based on OS
time.}

\item{tz}{String, timezone of original data (may be different from
default_tz).}
}
\value{
Original dataframe with a new "result" column
}
\description{
Examines two time columns, returns the greater or lesser value and inserts
  that value in a result column
}
\details{
Used in RoostArrivalDeparture(). Ensures that result is in the
  correct timezone.
}
