% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{LogisticByInflection}
\alias{LogisticByInflection}
\title{Logistic by inflection}
\usage{
LogisticByInflection(x, inflection, scale)
}
\arguments{
\item{x}{Numeric, x value}

\item{inflection}{Numeric, logistic function inflection point parameter}

\item{scale}{Numeric, logistic function scale parameter}
}
\value{
The y value
}
\description{
Returns the value of the logistic function using inflection and scale as
  input parameters
}
