% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NLLBeta}
\alias{NLLBeta}
\title{Negative Log-likelihood of a Beta distribution}
\usage{
NLLBeta(data, shape1, shape2)
}
\arguments{
\item{data}{Vector of data}

\item{shape1}{Numeric, shape1}

\item{shape2}{Numeriv, shape2}
}
\value{
NLL value
}
\description{
Beta Negative Log-Likelihood function
}
\details{
Used in BehaviorFit()
}
