% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NLLGamma}
\alias{NLLGamma}
\title{Negative Log-likelihood of a Gamma}
\usage{
NLLGamma(data, scale, shape)
}
\arguments{
\item{data}{Vector of data}

\item{scale}{Numeric, shape parameter}

\item{shape}{Numeric, scale parameter}
}
\value{
NLL of Gamma
}
\description{
Gamma Negative Log-Likelihood function
}
