% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NLLGenVonMises}
\alias{NLLGenVonMises}
\title{Negative log-likelihood of general von Mises}
\usage{
NLLGenVonMises(data, mu1, mu2, kappa1, kappa2)
}
\arguments{
\item{data}{Vector of data}

\item{mu1}{Numeric, primary direction paramter}

\item{mu2}{Numeric, secondary direction parameter}

\item{kappa1}{Numeric, non-negative parameter}

\item{kappa2}{Numeric, non-negative parameter}
}
\value{
NLL of general von Mises distribution
}
\description{
Negative log-likelihood of general von Mises function
}
\details{
Usage NLLGenVonMises(data, mu1, mu2, kappa1, kappa2)
}
