% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NLLPareto}
\alias{NLLPareto}
\title{Pareto negative Log-Likelihood function}
\usage{
NLLPareto(data, location, scale, shape)
}
\arguments{
\item{data}{Vector of data}

\item{location}{Numeric, location parameter ("left edge" of the probability
density)}

\item{scale}{Numeric, scale parameter}

\item{shape}{Numeric, shape parameter}
}
\value{
NLL of Pareto distribution
}
\description{
Negative Log-Likelihood of the Pareto function
}
\details{
uses the generalized Pareto distribution from the VGAM package
}
