% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NLLWeibull}
\alias{NLLWeibull}
\title{Negative log-likelihood of the Weibull function}
\usage{
NLLWeibull(data, shape, mu)
}
\arguments{
\item{data}{Vector, data}

\item{shape}{Numeric, shape parameter}

\item{mu}{Numeric, mu parameter}
}
\value{
NLL of Weibull distribution
}
\description{
Negative Log-Likelihood of the Weibull function based on shape and mu
  parameters.
}
\details{
used in DepartFit() and ArriveFit()
}
