% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{NonlinearRangeRescale}
\alias{NonlinearRangeRescale}
\title{NonlinearRangeRescale}
\usage{
NonlinearRangeRescale(x, min, max, lowBound, upBound, pars,
  movement_kernel)
}
\arguments{
\item{x}{value to rescale.}

\item{min}{minimum value of x on its original scale. Default is NULL.}

\item{max}{maximum value of x on its original scale. Default is NULL.}

\item{lowBound}{the minimum value of the desired scale. Default is 1.}

\item{upBound}{the maximum value of the desired scale. Default is NULL.}

\item{pars}{a list or numeric vector containing the scale and shape
parameters of a Gamma distribution in that order.}

\item{movement_kernel}{a raster or extent object used to calculate the value
for upBound.}

\item{negative}{a logical indicating whether or not the final rescaled value
should be multiplied by -1.}
}
\value{
rescaled x
}
\description{
Performs a nonlinear range rescale according the complement of a Gamma
  cumulative density function.
}
\details{
Based on code written by Javan Bauder and Kevin McGarigal. This
  function currently only uses a Gamma distribution. Arguments whose defaults
  are NULL are calculated internally based on the Gamma parameters provided
  and the extent of the movement_kernel raster or extent.
}
