% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{PlotBehaviorProportionBar}
\alias{PlotBehaviorProportionBar}
\title{Plots daily behavior proportions as bars}
\usage{
PlotBehaviorProportionBar(df, breaks, title)
}
\arguments{
\item{df}{Dataframe with "sex", "behavior", and "time_proportion" columns.}

\item{breaks}{Numeric, number of breaks in daily period.}

\item{title}{Character, main title of plot. Default is "Daily Behavior
Distributions".}
}
\value{
Facetted plot of behavior proportion over daily period.
}
\description{
Plots proportion of behavioral states during a day as a bar graph, with
  adjustable number of breaks.
}
\details{
Behavioral colors come from:
  "C:/Work/R/Data/BAEA/Models/Behavior_Colors.csv".
}
