% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotBetaPDF}
\alias{PlotBetaPDF}
\title{Plot Beta probability density function}
\usage{
PlotBetaCDF(shape1, shape2)
}
\arguments{
\item{shape1}{Numeric, shape1 parameter.}

\item{shape2}{Numeric, shape2 parameter.}

\item{color}{String, line color, default is "darkgreen".}
}
\value{
Plot of BetaPDF
}
\description{
Plots Beta probability density function using shape1 and shape2 parameters.
}
\details{
The x-axis is set to go to 500.
}
