% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotCauchyPDF}
\alias{PlotCauchyPDF}
\title{Plot Cauchy probability density function}
\usage{
PlotCauchyPDF(location = 5, scale = 5, xlim = NULL,
  color = "darkgreen")
}
\arguments{
\item{location}{Numeric, location. Default is 5.}

\item{scale}{Numeric, scale. Default is 1.}

\item{xlim}{Numeric, sets axis limits of plot. Default is: c(1st, 99th
quantile)}

\item{color}{String, line color. Default is "darkgreen".}
}
\value{
Plot of Cauchy PDF
}
\description{
Plot Cauchy probability density function using location and scale parameters
}
\details{
The x-axis is set to have 500 values
}
