% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PlotColorPalette}
\alias{PlotColorPalette}
\title{Plots selected color palettes}
\usage{
PlotColorPalette(pal, sel)
}
\arguments{
\item{pal}{Color palette, not in quotes (e.g. R_pal, SAGA_pal)}

\item{sel}{Vector, layer(s) in color palette to be displayed}
}
\value{
Plot of color palette(s)
}
\description{
Plots selected color palettes (e.g. R_pal, SAGA_pal) with indexes and names
}
\details{
Plot of color palette(s)
}
