% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PlotColorPie}
\alias{PlotColorPie}
\title{Plots colors in pie chart}
\usage{
PlotColorPie(pal, names = TRUE, radius = 1)
}
\arguments{
\item{pal}{the color palette (e.g. R_pal, SAGA_pal)}

\item{names}{Logical, whether or not to show name assigned to each color.
Default is TRUE.}

\item{radius}{Numeric, the pie is drawn centered in a square box whose sides
range from -1 to 1. If the character strings labeling the slices are long
it may be necessary to use a smaller radius. Default is 1.}
}
\value{
Plot of color pie chart
}
\description{
Plots selected colors from palettes in a pie chart
}
\details{
Must attach appropriate color palettes prior to running function.
}
