% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotGammaPDF}
\alias{PlotGammaPDF}
\title{Plots Gamma probability distribution function}
\usage{
PlotGammaPDF(shape = 1, scale = 1, col = "slateblue4", max_x = 25)
}
\arguments{
\item{shape}{Numeric, shape parameter. Default is 1.}

\item{scale}{Numeric, scale parameter. Default is 1.}

\item{col}{String, line color. Default is "slateblue4".}

\item{max_x}{Numeric, maximum valur on x scale. Default is 25.}
}
\value{
Plot of proability distribution
}
\description{
Plots Gamma probability distribution function using shape and scale
 parameters.
}
