% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotHistAndGenVonMises}
\alias{PlotHistAndGenVonMises}
\title{Plot histogram of data and generalized von Mises distribution}
\usage{
PlotHistAndGenVonMises (df, var, by, pars, bin_width,
  fit_gen_von_mises, fit_color, x_lab, title)
}
\arguments{
\item{df}{Dataframe of data}

\item{var}{Column name, variable to fit von Mises distribution}

\item{by}{Column name, optional, column name used to subset data, default
is NULL.}

\item{pars}{Parameter valuess, optional. A set of parameters used in place
of pars generated within the function. Default is NULL.}

\item{bin_width}{Numeric, bin size, default is: 15 degrees or 2*pi/24 radians}

\item{fit_gen_von_mises}{Logical, whether or not to fit and show
generalized von Mises distribution. Default is TRUE.}

\item{fit_color}{String, color used for von Mises fit line, quantiles, and
parameter value text. Default is "black"}

\item{x_lab}{String, name for x-axis, default is 'var'.}

\item{title}{String, title of plot, default is NULL.}
}
\value{
Plots of the data with a fitted von Mises distribution.
}
\description{
Plots a histogram of the data with an option to fit a generalized von Mises
  distribution
}
\details{
Automatically adjusts plot for degrees or radians input, but all
  parameter estimates are based on radians
}
