% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotHistAndNormal}
\alias{PlotHistAndNormal}
\title{Plots histogram and Normal distribution}
\usage{
PlotHistAndNormal(df, var, by, pars, xlim, bin_width, fit_normal,
  fit_color, x_lab, title, hold_axes, labels, lines)
}
\arguments{
\item{df}{Dataframe of data}

\item{var}{Column name of variable to fit Normal distribution}

\item{by}{Column name used to subset data, optional. Default is NULL}

\item{pars}{Parameters, optional. A set of parameters used in place of pars
generated within the function. Default is NULL.}

\item{xlim}{Numeric, x-value axis limits, if only one value is given, that
is used as the max limit. Default is NULL.}

\item{bin_width}{Numeric, bin size. Default is: x-value range/30}

\item{fit_normal}{Logical, whether or not to fit and show Normal
distribution. Default is TRUE.}

\item{fit_color}{String, color used for Normal fit line, quantiles, and
parameter value text. Default is "orangered".}

\item{x_lab}{String, label for x-axis. Default is variable name.}

\item{title}{String, main title name. Default is NULL.}

\item{hold_axes}{Logical, hold axes even if parameter distribution goes off
of the screen. Default is TRUE.}

\item{labels}{Logical, show quantile labels. Default is TRUE.}

\item{lines}{Logical, show quantile lines. Default is TRUE.}
}
\value{
Plots with a Normal distribution fitted for each 'by' factor
}
\description{
Plots a histogram of the data with an option to fit a Normal distribution.
}
