% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotHistAndPareto}
\alias{PlotHistAndPareto}
\title{Plots a histogram of the data and a Pareto distribution}
\usage{
PlotHistAndPareto(df, var, by, pars, xlim, bin_width, fit_pareto,
  fit_color, x_lab, title, hold_axes, emp_labels, emp_lines, labels, lines)
}
\arguments{
\item{df}{Dataframe of data}

\item{var}{Column name of variable to fit Pareto distribution}

\item{by}{Column name used to subset data, optional. Default is NULL.}

\item{pars}{Parameters used in place of pars generated within the function.
Optional. Default is NULL.}

\item{xlim}{Numeric, x-value limit. Default is NULL.}

\item{bin_width}{Numeric, bin size. Default is x-value range/30.}

\item{fit_pareto}{Logical, whether or not to fit and show Pareto
distribution. Default is TRUE.}

\item{fit_color}{String, color used for Pareto fit line, quantiles, and
parameter value text. Default is "orangered".}

\item{x_lab}{String, x-axis label. Default is NULL.}

\item{title}{String, title of plot. Default is NULL.}

\item{hold_axes}{Logical, hold axes even if parameter distribution goes off
of the screen. Default is TRUE.}

\item{emp_labels}{Logical, whether or not to show labels for empirical data
75th and  95th quantiles. Default is TRUE.}

\item{emp_lines}{Logical, whether or not to show lines for empirical data
75th and  95th quantiles. Default is TRUE.}

\item{labels}{Logical, show quantile labels. Default is TRUE.}

\item{lines}{Logical, show quantile lines. Default is TRUE.}
}
\value{
Plots with a Pareto distribution fitted for each 'by' factor
}
\description{
Plots a histogram of the data with an option to fit a Pareto distribution
}
