% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotHistAndWrappedCauchy}
\alias{PlotHistAndWrappedCauchy}
\title{Plots a histogram with a wrapped Cauchy distribution}
\usage{
PlotHistAndWrappedCauchy(df, var, by = NULL, pars = NULL,
  bin_width = NULL, fit_wrapped_cauchy = TRUE, fit_color = "black",
  x_lab = NULL, title = NULL)
}
\arguments{
\item{df}{Dataframe of data}

\item{var}{Column name of variable to fit wrapped Cauchy distribution}

\item{by}{Column name used to subset data, optional. Default is NULL.}

\item{pars}{A set of parameters used in place of pars generated within the
functions, optional. Default is NULL.}

\item{bin_width}{Numeric, bin size. Default is 15 degrees or 2*pi/24 radians.}

\item{fit_wrapped_cauchy}{Logical, whether or not to fit and show wrapped
Cauchy distribution. Default is TRUE.}

\item{fit_color}{String, color used for Pareto fit line, quantiles, and
parameter value text. Default is "black".}

\item{x_lab}{String, name for x-axis. Default is 'var'.}

\item{title}{String, name for title.}
}
\value{
A plot of the data with a fitted wrapped Cauchy distribution.
}
\description{
Plots a histogram of the data with an option to fit a wrapped Cauchy
  distribution
}
\details{
Usage: PlotHistAndWrappedCauchy(df, var, by, pars, bin_width,
  fit_wrapped_cauchy, fit_color, x_lab, title)

Automatically adjusts plot for degrees or radians input, but all
  parameter estimates are based on radians.
}
