% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PlotHistogramByVar}
\alias{PlotHistogramByVar}
\title{Plot histogram by grouping variable}
\usage{
PlotHistogramByVar(df, var, id = "id")
}
\arguments{
\item{df}{Dataframe of data}

\item{var}{Column name with data to create histogram}

\item{id}{Column name of unique identified. Default is "id".}
}
\value{
A histogram plot
}
\description{
Plots a histogram of a variable by id of a grouping variable
}
