% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotLogisticInflectionPDF}
\alias{PlotLogisticInflectionPDF}
\title{Plot logistic probability density function}
\usage{
PlotLogisticInflectionPDF(inflection, scale, range, col)
}
\arguments{
\item{inflection}{Numeric, inflection parameter.}

\item{scale}{Numeric, scale parameter.}

\item{range}{Vector for x-scale range, default is (0,50).}

\item{col}{String, color palette, e.g. "jet2.col(100)".}
}
\value{
Plot of probability distribution
}
\description{
Plot logistic probability density function using a logistic function with
  inflection and scale parameters
}
