% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotLogisticPDF}
\alias{PlotLogisticPDF}
\title{Plot Logistic probability distribution function}
\usage{
PlotLogisticPDF(location, scale, range, col)
}
\arguments{
\item{location}{Numeric, location parameter}

\item{scale}{Numeric, location parameter}

\item{range}{Vector for x-scale range, default is (0,50)}

\item{col}{String, color palette, e.g. "jet2.col(100)"}
}
\value{
Plot of probability distribution
}
\description{
Plot Logistic Probability Distribution Function using logistic function with
  location and scale parameters
}
