% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotParetoPDF}
\alias{PlotParetoPDF}
\title{Plot Pareto probability density function}
\usage{
PlotParetoPDF(location = 1, scale = 1, shape = 1, xlim = NULL,
  color = "darkgreen")
}
\arguments{
\item{location}{Numeric, location parameter, default is 1}

\item{scale}{Numeric, scale parameter, default is 1.}

\item{shape}{Numeric, shape parameter, default is 1.}

\item{xlim}{Numeric, sets plot x lim max, default is 99th percentile.}

\item{color}{}
}
\value{
Plot of Parteo probability density function
}
\description{
Plot Pareto probability density function using the location, scale, and
  shape parameters
}
\details{
x-axis is set to have 500 values. Commented out section: 75th and
  95th percentiles are shown with dashed red lines.
}
