% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotParsBeta}
\alias{PlotParsBeta}
\title{Plot Beta probability density functions}
\usage{
PlotParsBeta(pars, length)
}
\arguments{
\item{pars}{List, parameter values.}

\item{length}{Integer, number of x values used to plot lines, default is 100.}
}
\value{
Plot of BetaPDF with the parameter values of parameter list
}
\description{
Plot Beta probability density functions from list of parameter values
}
\details{
Automatically adds 1 to the length, since 0 is included
}
