% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{PlotProbabilityRaster}
\alias{PlotProbabilityRaster}
\title{Plot 3D raster of selected cells given a probability raster and sample size}
\usage{
PlotProbabilityRaster(prob_raster, sample_n = 5000)
}
\arguments{
\item{prob_raster}{Raster}

\item{sample_n}{integer sample size of selected cells from probability raster}
}
\value{
Raster
}
\description{
Plot 3D raster of selected cells given a probability raster and sample size
}
