% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{PlotRoostHistogram}
\alias{PlotRoostHistogram}
\title{Plots roost behavior probability distribution}
\usage{
PlotRoostHistogram(df, pars)
}
\arguments{
\item{df}{Dataframe of location data}

\item{pars}{List of simulation parameters with male/female, arrive/depart,
shape/scale}

\item{binwidth}{Numeric, bin size. Default is 15.}
}
\value{
Facetted plot with roost data and fitted weibull distributions
}
\description{
Plots roost times and fitted Weibull probability distributions
}
\details{
Weibull probability function extends 15 min past last time.
}
