% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{PlotStepLengths}
\alias{PlotStepLengths}
\title{Plot step lengths histogram}
\usage{
PlotStepLengths(df, id = "id", xlim = NULL, bin_width = NULL)
}
\arguments{
\item{df}{Dataframe with flights.}

\item{id}{String, column name of unique identifier. Default = "id".}

\item{xlim}{Numeric, x-value limit. Default is NULL.}

\item{bin_width}{Numeric, bin size. Default is: x-value range/30}
}
\value{
A plot of step-length distances
}
\description{
Plots a histogram of the step-lengths for each bird
}
