% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotTwoLogisticCDF}
\alias{PlotTwoLogisticCDF}
\title{Plot two logistic probability density functions simultaneously}
\usage{
PlotTwoLogisticPDF(location_1, scale_1, location_2, scale_2, x_label,
  x_max)
}
\arguments{
\item{location_1}{Numeric, location parameter for distribution 1.}

\item{scale_1}{Numeric, scale parameter for distribution 1.}

\item{location_2}{Numeric, location parameter for distribution 2.}

\item{scale_2}{Numeric, scale parameter for distribution 2.}

\item{x_label}{String, label for x-axis, default if "X Value".}

\item{x_max}{Numeric, maximum value on x scale, default is 1000.}
}
\value{
Plot of probability distribution
}
\description{
Plot logistic probability distribution function using location and scale
  parameters.
}
