% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotTwoLogisticInflectionPDF}
\alias{PlotTwoLogisticInflectionPDF}
\title{Plot two logistic probability density function}
\usage{
PlotTwoLogisticInflectionPDF(
  inflection_1,
  scale_1,
  inflection_2,
  scale_2,
  x_label = "X Value",
  log1_range = c(0, 45),
  log2_range = c(0, 45),
  xlab_range = c(0, 45)
)
}
\arguments{
\item{inflection_1}{Numeric, inflection parameter for distribution 1.}

\item{scale_1}{Numeric, scale parameter for distribution 1.}

\item{inflection_2}{Numeric, inflection parameter for distribution 2.}

\item{scale_2}{Numeric, scale parameter for distribution 2.}

\item{x_label}{String, label for x-axis, default if "X Value".}

\item{log1_range}{Vector, log1 value range. Default = c(0, 45).}

\item{log2_range}{Vector, log2 value range. Default = c(0, 45).}

\item{xlab_range}{Vector, x value range. Default = c(0, 45).}
}
\value{
Plot of probability distributions.
}
\description{
Plot two logistic probability density function (using inflection and scale
  parameters)
}
