% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{PlotWeibullPDF}
\alias{PlotWeibullPDF}
\title{Plot Weibull cumulative distribution function}
\usage{
PlotWeibullCDF(shape, scale, max_x)
}
\arguments{
\item{shape}{Numeric, Weibull shape, default is 1.}

\item{scale}{Numeric, Weibull scale, default is 1.}

\item{max_x}{Numeric, maximum value on x scale, default is 250.}
}
\value{
Plot of probability distribution
}
\description{
Plot Weibull cumulative distribution function using shape and scale
  parameters
}
