% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RemoveExcept}
\alias{RemoveExcept}
\title{Removes all objects in global environment except selected ones.}
\usage{
RemoveExcept(object = object)
}
\arguments{
\item{object}{Object(s) to keep in environment, e.g., c("df1", "vec1").}
}
\value{
New global environment
}
\description{
Removes all objects in global environment except ones included in the
  argument in quotes, e.g., c("df1", "vec1").
}
