% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ReplaceFilesText}
\alias{ReplaceFilesText}
\title{Replace text in files}
\usage{
ReplaceFilesText(files, text, replace)
}
\arguments{
\item{files}{List of filenames}

\item{text}{String of text to search for in files}

\item{replace}{String for replacement in files}
}
\value{
Files with replaced text
}
\description{
Replaces a text string within files
}
