% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{SaveGGPlot}
\alias{SaveGGPlot}
\title{Save last printed ggplot}
\usage{
SaveGGPlot(filename, path, width, height, units, dpi, bg)
}
\arguments{
\item{filename}{String, file name/filename of plot}

\item{path}{String, path to save plot to (if you just want to set path and
not filename). If NULL(default), reverts to working directory}

\item{width}{Numeric, width. Default is 10}

\item{height}{Numeric, height. Default is 7.5}

\item{units}{String, units for width and height when either one is
explicitly specified (in, cm, or mm). Default is "in".}

\item{dpi}{Integer, dpi to use for raster graphics. Default is 300.}

\item{bg}{String, background color. Default is "white"}
}
\value{
Saves a png file of the last displayed plot
}
\description{
A wrapper function for ggsave()
}
\details{
Default output format is set for PowerPoint presentations
}
