% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{SavePlot}
\alias{SavePlot}
\title{Saves last plot}
\usage{
SavePlot(filename, path, width, height, units, dpi)
}
\arguments{
\item{filename}{String, file name/filename of plot.}

\item{path}{String, path to save plot to (if you just want to set path and
not filename). If NULL(default), uses working directory.}

\item{width}{Numeric, width. Default is 10.}

\item{height}{Numeric, height. Default is 7.5.}

\item{units}{Numeric, units for image width and height when either one is
explicitly specified (in, cm, or mm), default is "in".}

\item{dpi}{Integer, dpi to use for plotting. Default is 300.}
}
\value{
Saves a file of the last displayed plot.
}
\description{
A function for saving the last displayed plot as a png, jpeg, or pdf.
}
\examples{
Default output format width and height are set for PowerPoint
  presentations

}
