% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{SimplifySimSpatialBAEA}
\alias{SimplifySimSpatialBAEA}
\title{Simplify the sim object so that the sim's object size is smaller}
\usage{
SimplifySimSpatialBAEA(sim)
}
\arguments{
\item{sim}{}
}
\value{

}
\description{
Simplify the sim object so that the sim's object size is smaller
}
