% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{SummarizeFitdist}
\alias{SummarizeFitdist}
\title{Summarizes a 'fitdist' object}
\usage{
SummarizeFistdist(fits)
}
\arguments{
\item{fits}{'fitdist' object.}
}
\value{
Dataframe
}
\description{
Summarizes a 'fitdist' object (from 'fitdistrplus' package), made to be used
  as a table (e.g. xtable)
}
