% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{SummarizeSE}
\alias{SummarizeSE}
\title{Summarize data and standard error}
\usage{
SummarizeSE(df, var, groups, na_rm, conf_int)
}
\arguments{
\item{data}{Dataframe}

\item{var}{Column name that contains the variable to be summarized}

\item{by}{Column name of vector containing grouping variables, Default is
NULL.}

\item{na_rm}{Logical that indicates whether to ignore NAs.}

\item{conf_interval}{Numeric, the percent range of the confidence interval.
Default is .95)}
}
\value{
Dataframe
}
\description{
Summarizes data with count, mean, standard deviation, standard error of the
   mean, and confidence interval
}
\details{
Based on example function from "Cookbook for R" website.
}
