% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ThemeBlack}
\alias{ThemeBlack}
\title{Black theme for ggplot}
\usage{
ThemeBlack(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Numeric, font base size. Default is 12.}

\item{base_family}{String, font name. Default is "".}
}
\value{
A ggplot theme
}
\description{
A function that creates a black background/white text theme for ggplot
}
\details{
Original script is from: \url{https://gist.github.com/jslefche}
}
