% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{TocMsg}
\alias{TocMsg}
\title{Message for toc of tictoc}
\usage{
TocMsg(tic, toc, msg, info)
}
\arguments{
\item{tic}{tictoc object}

\item{toc}{tictoc object}

\item{msg}{character, message}

\item{info}{character, other information}
}
\description{
Message for toc of tictoc
}
