% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{UpdateAgentStatesBAEA}
\alias{UpdateAgentStatesBAEA}
\title{Update agent_state object in BAEA IBM}
\usage{
UpdateAgentStatesBAEA(agent_states = NULL, sim = sim, init = FALSE)
}
\arguments{
\item{init}{}
}
\value{

}
\description{
Update agent_state object in BAEA IBM
}
