% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{UpdateWeeklyData}
\alias{UpdateWeeklyData}
\title{Updates weekly baea data and send emails}
\usage{
UpdateWeeklyData(data = data, download_recent = FALSE, send_email = TRUE,
  date = now(), to = c("erynn.call@maine.gov", "charlie.todd@maine.gov"))
}
\arguments{
\item{data}{Dataframe of deployed data.}

\item{download_recent}{Logical, whether or not to download recent GPS data,
default is FALSE}

\item{send_email}{Logical, whether or not to send an email, default is TRUE.}

\item{date}{Date within week to update, default is now().}

\item{to}{String, recipients of email, default: c("erynn.call@maine.gov",
"charlie.todd@maine.gov").}
}
\value{
Creates files in "C:/Work/R/Data/BAEA/Telemetry" and
}
\description{
Compiles and manages csv files and kml files in order to create individual
  .csv and .kml files, weekly .kml files, and send an email with weekly .kml
  file.
}
\details{
Specific to my Bald Eagle Project.
}
