% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{UpdateWeeklyKMLFiles}
\alias{UpdateWeeklyKMLFiles}
\title{Updates weekly .kml files}
\usage{
UpdateWeeklyKMLFiles(data, start, end, update_gdrive = TRUE)
}
\arguments{
\item{data}{Dataframe of location data.}

\item{start}{String, start date.}

\item{end}{String, end date.}

\item{update_gdrive}{Logical, whether to update the files on the GDrive.
Default is TRUE}
}
\value{
Creates .kml files in "C:/Work/R/Data/BAEA/Telemetry" and
  "C:/Users/Blake/Google Drive/BAEA Project/Telemetry Data" folders.
}
\description{
Updates weekly KML files on local and GDrive folders. All weekly intervals
  are set to be from Sunday to Saturday, inclusive. All start and end dates
  will be adjusted to the beginning or end, respectively, of that week.
}
