% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{AddCruiseBehavior}
\alias{AddCruiseBehavior}
\title{Adds cruise behavior}
\usage{
AddCruiseBehavior(df, alt_abv_grd, min_speed)
}
\arguments{
\item{df}{dataframe}

\item{min_agl}{distance (in meters) above ground}

\item{min_speed}{minimum speed of bird}
}
\value{
dataframe with behavior column that has "cruise"
}
\description{
Finds cruise behavior that meet given threshold parameters
}
\details{
should run AddNestBehavior, AddRoostBehavior, and AddCruiseBehavior
  prior to this function
}
