% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{ExtractTransitionProbabilities}
\alias{ExtractTransitionProbabilities}
\title{Extract 'momentuHMM' transition matrix probabilities}
\usage{
ExtractTransitionProbabilities(x, alpha = 0.95)
}
\arguments{
\item{x}{=  object}

\item{alpha}{= p-value for CI}
}
\value{
a tibble
}
\description{
Extract 'momentuHMM' transition matrix probabilities
}
\details{
This function is based on momentuHMM:::plot function. There may be
    a lot of extraneous code because it was difficult to extract only the
    necessary code for creating the transition matrix.
}
