% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{PlotBehaviorProportionLine}
\alias{PlotBehaviorProportionLine}
\title{Plots daily behavior proportions as lines}
\usage{
PlotBehaviorProportionLine(df, breaks)
}
\arguments{
\item{df}{Dataframe with "sex", "behavior", and "time_proportion" columns.}

\item{breaks}{Numeric, number of breaks in daily period.}
}
\value{
Facetted plot of behavior proportion over daily period.
}
\description{
Plots proportion of behavioral states during a day as a line, with
  adjustable number of breaks.
}
