% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{SummarizeFitDist}
\alias{SummarizeFitDist}
\title{Summarizes a 'fitdist' object}
\usage{
SummarizeFistDist(fits)
}
\arguments{
\item{fits}{'fitdist' object.}

\item{orientation}{numeric, 1 = 'rowwise' or 2 = 'colwise', method for how
distributions' information are arranged}
}
\value{
tibble
}
\description{
Summarizes a 'fitdist' object (from 'fitdistrplus' package), made to be used
  as a table (e.g. xtable)
}
